issue-1164: splitting endpoint_proxy into server and client libs (client lib is currently empty), making a tiny fix in server.cpp
