+++ cloud/blockstore/libs/endpoint_proxy/server/ya.make
@@ -1,0 +1,17 @@
+LIBRARY()
+
+SRCS(
+    bootstrap.cpp
+    options.cpp
+    server.cpp
+)
+
+PEERDIR(
+    cloud/blockstore/libs/daemon/common
+
+    library/cpp/getopt/small
+
+    contrib/ydb/library/actors/util
+)
+
+END()
